Use double precision

+++ src/EquationSolver.cs
@@ -5,21 +5,21 @@ namespace Intersect {
 
     public static class EquationSolver
     {
-        public static float SolveLinear(float a, float b) {
+        public static double SolveLinear(double a, double b) {
             return -b / a;
         }
 
-        public static IList<float> SolveQuadratic(float a, float b, float c) {
-            List<float> output = new List<float>();
-            if (a == 0f) {
+        public static IList<double> SolveQuadratic(double a, double b, double c) {
+            List<double> output = new List<double>();
+            if (a == 0d) {
                 output.Add(SolveLinear(b, c));
                 return output;
             }
-            var discriminant = b * b - 4.0f * a * c;
-            var divider = 2.0f * a;
-            if (discriminant == 0.0f) {
+            var discriminant = b * b - 4d * a * c;
+            var divider = 2d * a;
+            if (discriminant == 0d) {
                 output.Add(-b / divider);
-            } else if (discriminant > 0.0f) {
+            } else if (discriminant > 0d) {
                 var root = (float)Math.Sqrt(discriminant);
                 output.Add((-b - root) / divider);
                 output.Add((-b + root) / divider);
@@ -27,10 +27,10 @@ namespace Intersect {
             return output;
         }
 
-        public static IList<float> SolveQubic(float a, float b, float c, float d) {
+        public static IList<double> SolveQubic(double a, double b, double c, double d) {
             // Taken from: https://stackoverflow.com/questions/49671821/implementing-cardano-method-for-a-cubic-equation-solution
             // And: https://en.wikipedia.org/wiki/Cubic_equation#Depressed_cubic
-            List<float> output = new List<float>();
+            List<double> output = new List<double>();
             if (a == 0f) {
                 return SolveQuadratic(b, c, d);
             }
@@ -43,22 +43,22 @@ namespace Intersect {
             if (cubeDiscr > 0) {
                 double u = CubicRoot(-q / 2d + Math.Sqrt(cubeDiscr));
                 double v = CubicRoot(-q / 2d - Math.Sqrt(cubeDiscr));
-                output.Add((float)(u + v - (a1 / 3d)));
+                output.Add(u + v - (a1 / 3d));
             }
             else if (cubeDiscr == 0) {
                 double u = CubicRoot(-q / 2d);
                 if (q != 0d) {
                     // Prevent duplicate result.
-                    output.Add((float)(2d * u - (a1 / 3d)));
+                    output.Add(2d * u - (a1 / 3d));
                 }
-                output.Add((float)(-u - (a1 / 3d)));
+                output.Add(-u - (a1 / 3d));
             }
             else if (cubeDiscr < 0) {
                 double u = 2d * Math.Sqrt(p / -3d);
                 double theta = Math.Acos(3d * q * Math.Sqrt(-3d / p) / (2d * p)) / 3d;
                 for(int i = 0; i < 3; i++) {
                     double t = u * Math.Cos(theta - 2d * Math.PI * i / 3d);
-                    output.Add((float)(t - a1 / 3d));
+                    output.Add(t - a1 / 3d);
                 }
             }
             return output;
